HW 3 after review

+++ src/main/java/ru/javawebinar/topjava/repository/jdbc/JdbcMealRepository.java
@@ -10,12 +10,9 @@ import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
-import ru.javawebinar.topjava.util.Util;
-
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.function.Predicate;
-import java.util.stream.Collectors;
+
 
 @Repository
 public class JdbcMealRepository implements MealRepository {
@@ -37,13 +34,12 @@ public class JdbcMealRepository implements MealRepository {
         this.namedParameterJdbcTemplate = namedParameterJdbcTemplate;
     }
 
-
     @Override
     public Meal save(Meal meal, int userId) {
         MapSqlParameterSource map = new MapSqlParameterSource()
-                .addValue("user_id", userId)
                 .addValue("id", meal.getId())
-                .addValue("dateTime", meal.getDateTime())
+                .addValue("user_id", userId)
+                .addValue("date_time", meal.getDateTime())
                 .addValue("description", meal.getDescription())
                 .addValue("calories", meal.getCalories());
 
@@ -51,8 +47,8 @@ public class JdbcMealRepository implements MealRepository {
             Number newKey = insertMeal.executeAndReturnKey(map);
             meal.setId(newKey.intValue());
         } else if (namedParameterJdbcTemplate.update(
-                "UPDATE meals SET dateTime=:dateTime, description=:description, calories=:calories" +
-                        " WHERE id=:id", map) == 0) {
+                "UPDATE meals SET date_time=:date_time, description=:description, calories=:calories" +
+                        " WHERE user_id=:user_id AND id=:id", map) == 0) {
             return null;
         }
         return meal;
@@ -60,7 +56,7 @@ public class JdbcMealRepository implements MealRepository {
 
     @Override
     public boolean delete(int id, int userId) {
-        return jdbcTemplate.update(" DELETE FROM meals WHERE  id=?", id) != 0;
+        return jdbcTemplate.update(" DELETE FROM meals WHERE user_id=? AND id=?", userId, id) != 0;
     }
 
     @Override
@@ -71,19 +67,12 @@ public class JdbcMealRepository implements MealRepository {
 
     @Override
     public List<Meal> getAll(int userId) {
-        return filterList(meal -> true, userId);
+        return jdbcTemplate.query(" SELECT * FROM meals WHERE user_id=? ORDER BY date_time DESC ", ROW_MAPPER, userId);
     }
 
     @Override
     public List<Meal> getBetweenHalfOpen(LocalDateTime startDateTime, LocalDateTime endDateTime, int userId) {
-        return filterList(meal -> Util.isBetweenHalfOpen(meal.getDateTime(), startDateTime, endDateTime), userId);
-    }
-
-    public List<Meal> filterList(Predicate<Meal> filter, int userId) {
-        List<Meal> meals = jdbcTemplate.query(" SELECT * FROM meals WHERE user_id=?", ROW_MAPPER, userId);
-        return meals.stream()
-                .filter(filter)
-                .collect(Collectors.toList());
+        return jdbcTemplate.query(" SELECT * FROM meals WHERE user_id=? AND date_time   BETWEEN ? AND ? ORDER BY date_time DESC ", ROW_MAPPER, userId, startDateTime, endDateTime);
     }
 }
